% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.sandwichboost.R
\name{print.sandwichboost}
\alias{print.sandwichboost}
\title{Print for a sandwichboost fitted object}
\usage{
\method{print}{sandwichboost}(x, ...)
}
\arguments{
\item{x}{a fitted \code{sandwich} object fitted by \code{PLR_sandboost()}.}

\item{...}{additional arguments}
}
\description{
This is a method that prints a useful summary of aspects of a \code{sandwich} object fitted by \code{PLR_sandboost()}.
}
