% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R, R/sandwich.boost-package.R
\docType{package}
\name{sandwich.boost}
\alias{sandwich.boost}
\title{Sandwich Boosting: A package for performing sandwich boosting for the PLR model}
\description{
To learn more about sandwich.boost, start with the vignette:
`browseVignettes(package = "sandwich.boost")`

To learn more about sandwich.boost, start with the vignette:
`browseVignettes(package = "sandwich.boost")`
}
