% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.sandwichboost.R
\name{summary.sandwichboost}
\alias{summary.sandwichboost}
\title{Summary for a sandwichboost fitted object}
\usage{
\method{summary}{sandwichboost}(object, ...)
}
\arguments{
\item{object}{a fitted \code{sandwich} object fitted by \code{PLR_sandboost()}.}

\item{...}{additional arguments}
}
\description{
This is a method that prints a \code{sandwich} object fitted by \code{PLR_sandboost()}.
}
